+++ Diagnostics/Assets/Turandot/Cues/Turandot.Cues.Message.cs
@@ -2,6 +2,8 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.ComponentModel.Design;
+using System.Drawing.Design;
 using System.Xml.Serialization;
 
 using Newtonsoft.Json;
@@ -16,6 +18,7 @@ namespace Turandot.Cues
         public enum Icon { None, Right, Wrong, Warning}
 
         [Category("Appearance")]
+        [Editor(typeof(MultilineStringEditor), typeof(UITypeEditor))]
         public string Text { get; set; }
         private bool ShouldSerializeText() { return false; }